Split OpenAI chat internals
