+++ apps/web/src/server/chat/openai/tooling/tools.ts
